% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{compose}
\alias{compose}
\title{Compose a GgMagnify object into a single ggplot}
\usage{
compose(x, plot = x$plot)
}
\arguments{
\item{x}{A GgMagnify object.}

\item{plot}{A ggplot object, by default the original plot.}
}
\value{
A ggplot
}
\description{
Use this when you want to save the result of \code{\link[=ggmagnify]{ggmagnify()}} to a file.
}
\examples{
library(ggplot2)
ggp <- ggplot(ChickWeight, aes(Time, weight, group = Chick, color = Diet)) +
  geom_line()

ggm <- ggmagnify(ggp,
  xlim = c(0, 5), ylim = c(30, 80),
  inset_xlim = c(0, 15), inset_ylim = c(200, 350))
compose(ggm)
\dontrun{
  ggsave("filename.png", compose(ggm))
}
}
