% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{grob_where}
\alias{grob_where}
\title{Create a grob from a subset of sf data}
\usage{
grob_where(where, sf, crs = NULL)
}
\arguments{
\item{where}{Logical condition to be evaluated in \code{sf}.}

\item{sf}{An object of class \link[sf:sf]{sf}.}

\item{crs}{Optional \link[sf:st_crs]{coordinate reference system}.}
}
\value{
A \code{\link[grid:grid.grob]{grid::grob()}} which you can pass to \code{from}.
}
\description{
Create a grob from a subset of sf data
}
\examples{
library(ggplot2)
if (requireNamespace("sf", quietly = TRUE) &&
  requireNamespace("maps", quietly = TRUE)) {
  usa <- sf::st_as_sf(maps::map("state", fill=TRUE, plot = FALSE))
  texas <- grob_where(ID == "texas", usa, crs = sf::st_crs(4326))
  ggplot(usa) + coord_sf(crs = sf::st_crs(4326)) + geom_sf() +
    geom_magnify(from = texas, to = c(-90, -70, 25, 35), colour = "red",
    aspect = "fixed", expand = 0)
}
}
