% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmagnify.R
\name{inset_blanks}
\alias{inset_blanks}
\title{Default elements to blank in the ggmagnify inset}
\usage{
inset_blanks(..., axes)
}
\arguments{
\item{...}{Character vector of extra elements to blank.}

\item{axes}{Logical. \code{TRUE} if the inset will include axes.}
}
\value{
A character vector.
}
\description{
Default elements to blank in the ggmagnify inset
}
