% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{+.GgMagnify}
\alias{+.GgMagnify}
\title{Deprecated}
\usage{
\method{+}{GgMagnify}(e1, e2)
}
\arguments{
\item{e1}{A GgMagnify object}

\item{e2}{A layer, theme, etc.}
}
\value{
The modified object
}
\description{
Deprecated
}
