% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmagnify.R
\name{print.GgMagnify}
\alias{print.GgMagnify}
\title{Deprecated.}
\usage{
\method{print}{GgMagnify}(x, ...)
}
\arguments{
\item{x}{A GgMagnify object.}

\item{...}{Passed on to \code{\link[ggplot2:print.ggplot]{ggplot2::print.ggplot()}}.}
}
\value{
The GgMagnify object, invisibly.
}
\description{
Deprecated.
}
