% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-magnify-tile.R
\docType{data}
\name{StatMagnifyTile}
\alias{StatMagnifyTile}
\title{#' @rdname GeomMagnify
#' @format NULL
#' @usage NULL
#' @export
GeomMagnifyTile <- ggproto("GeomMagnifyTile", GeomMagnify,
required_aes = c("x", "y", "height", "width",
"to_x", "to_y", "to_height", "to_width"),}
\format{
An object of class \code{StatMagnifyTile} (inherits from \code{StatMagnify}, \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatMagnifyTile
}
\description{
draw_panel = function(self, ...) {
ggproto_parent(GeomMagnify, self)$draw_panel(...)
}
)
}
\keyword{datasets}
