% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{compose}
\alias{compose}
\title{Compose a GgMagnify object into a single ggplot}
\usage{
compose(x, plot = x$plot)
}
\arguments{
\item{x}{A GgMagnify object.}

\item{plot}{A ggplot object, by default the original plot.}
}
\value{
A ggplot
}
\description{
Compose a GgMagnify object into a single ggplot
}
