% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{compose}
\alias{compose}
\title{Compose a GgMagnify object into a single ggplot}
\usage{
compose(x)
}
\arguments{
\item{x}{A GgMagnify object}
}
\value{
A ggplot
}
\description{
Compose a GgMagnify object into a single ggplot
}
