% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmagnify.R
\name{compose}
\alias{compose}
\title{Compose a GgMagnify object with its ggplot}
\usage{
compose(x, plot)
}
\arguments{
\item{x}{A GgMagnify object.}

\item{plot}{A ggplot object.}
}
\value{
The modified plot.
}
\description{
Compose a GgMagnify object with its ggplot
}
