% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-magnify.R
\name{geom_magnify}
\alias{geom_magnify}
\title{Magnified inset of a plot}
\usage{
geom_magnify(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  magnify,
  expand = TRUE,
  proj = c("facing", "corresponding", "single"),
  recalculate = FALSE,
  linetype = 1,
  target.linetype = linetype,
  proj.linetype = 2,
  inset.linetype = linetype,
  alpha = 1,
  linewidth = 0.5,
  shape = "rect",
  plot = NULL,
  na.rm = FALSE,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{magnify}{Numeric. How much to magnify the target area? Give a length-2
vector for separate x- and y-magnification.}

\item{expand}{Logical. Expand the limits of the target area within the inset
plot?}

\item{proj}{String. What style of projection lines to draw? \code{"facing"} (the
default), \code{"corresponding"} or \code{"single"}. Can be abbreviated. See below.}

\item{recalculate}{Logical. If \code{TRUE}, use \code{\link[ggplot2:lims]{ggplot2::lims()}} to replot the
target area. Statistics, e.g. smoothing lines, will be recalculated using
only that subset of the data. If \code{FALSE}, use \code{\link[ggplot2:coord_cartesian]{ggplot2::coord_cartesian()}}
to replot the target area, keeping all the data.}

\item{linetype, alpha, linewidth}{Linetype, alpha and linewidth for borders
and projection lines. Set \code{linetype = 0} for no lines.}

\item{target.linetype}{Linetype for border round the target area.}

\item{proj.linetype}{Linetype for projection lines.}

\item{inset.linetype}{Linetype for border round the inset.}

\item{shape}{Not yet implemented.}

\item{plot}{Ggplot object to plot in the inset. If \code{NULL}, defaults to the
ggplot object to which \code{geom_magnify()} is added.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
\code{geom_magnify()} creates a magnified inset of part of a plot. The target
area is centred at \verb{(x,y)} with \code{width} and \code{height}. The inset is centred
at \verb{(to_x,to_y)}, with \code{width} and \code{height} multiplied by \code{magnify}. Optional
borders are drawn around the target and inset, along with projection lines
from one to the other.
}
\section{Aesthetics}{

geom_magnify understand the following aesthetics (required aesthetics are in
bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{width}
\item \strong{height}
\item \strong{to_x}
\item \strong{to_y}
\item colour
}
}

\examples{
library(ggplot2)
ggp <- ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
         geom_point() + xlim(c(2, 6))
ggp
ggp + geom_magnify(x = 3, width = 1, y = 6.5, height = 1,
                   to_x = 5, to_y = 5, magnify = 1.5)
ggp +
      facet_wrap(vars(Species)) +
      geom_magnify(x = 3, width = 1, y = 6.5, height = 1,
                   to_x = 5, to_y = 5, magnify = 1.5)
}
