% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-magnify.R
\name{geom_magnify}
\alias{geom_magnify}
\title{#' @export
StatMagnify <- ggproto("StatMagnify", Stat,
compute_layer = function (self, data, scales, ...) {
browser()
data
}
)}
\usage{
geom_magnify(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE
)
}
\description{
#' @export
StatMagnify <- ggproto("StatMagnify", Stat,
compute_layer = function (self, data, scales, ...) {
browser()
data
}
)
}
