% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-magnify.R, R/geom-magnify-tile.R
\name{geom_magnify}
\alias{geom_magnify}
\alias{geom_magnify_tile}
\title{Create a magnified inset of a plot}
\usage{
geom_magnify(
  mapping = NULL,
  data = NULL,
  stat = StatMagnify,
  position = "identity",
  ...,
  shape = c("rect", "ellipse", "outline"),
  expand = 0.1,
  aspect = c("free", "fixed"),
  axes = "",
  proj = c("facing", "corresponding", "single"),
  shadow = FALSE,
  corners = 0,
  colour = "black",
  linetype = 1,
  target.linetype = linetype,
  inset.linetype = linetype,
  proj.linetype = 2,
  alpha = 1,
  linewidth = 0.4,
  plot = NULL,
  shadow.args = list(sigma = 5, colour = "grey40", x_offset = 5, y_offset = 5),
  recompute = FALSE,
  scale.inset = 1,
  proj.combine = TRUE,
  na.rm = FALSE,
  inherit.aes = TRUE
)

geom_magnify_tile(
  mapping = NULL,
  data = NULL,
  stat = StatMagnifyTile,
  position = "identity",
  ...,
  shape = c("rect", "ellipse", "outline"),
  expand = 0.1,
  aspect = c("free", "fixed"),
  axes = "",
  proj = "facing",
  shadow = FALSE,
  corners = 0,
  colour = "black",
  linetype = 1,
  target.linetype = linetype,
  inset.linetype = linetype,
  proj.linetype = 2,
  alpha = 1,
  linewidth = 0.4,
  plot = NULL,
  shadow.args = list(sigma = 5, colour = "grey40", x_offset = 5, y_offset = 5),
  recompute = FALSE,
  scale.inset = 1,
  proj.combine = TRUE,
  na.rm = FALSE,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm}{See e.g. \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{shape}{Shape of the area to be magnified. \code{"rect"} for a rectangle.
\code{"ellipse"} for an ellipse. \code{"outline"} for the convex hull of points in the
target area, or for map polygons.}

\item{expand}{Number. Expand the target area and inset proportionally
by this amount.}

\item{aspect}{String. \code{"fixed"} to fix the aspect ratio (overrides \code{ymax}).}

\item{axes}{String. Which axes to plot in the inset? \code{""}, \code{"x"}, \code{"y"} or
\code{"xy"}.}

\item{proj}{String. What style of projection lines to draw? \code{"facing"} (the
default), \code{"corresponding"} or \code{"single"}. Can be abbreviated. See below.}

\item{shadow}{Logical. Draw a shadow behind the inset plot? Requires the
"ggfx" package.}

\item{corners}{Numeric between 0 and 1. Radius of rounded corners for the
target area and inset. Only used if \code{shape} is \code{"rect"}. 0.1 is a good
starting value.}

\item{linetype, colour, alpha, linewidth}{Linetype, colour, alpha and linewidth
for borders and projection lines.}

\item{target.linetype, inset.linetype, proj.linetype}{Linetypes
for specific components. Set to \code{0} for no lines.}

\item{plot}{Ggplot object to plot in the inset. If \code{NULL}, defaults to the
ggplot object to which \code{geom_magnify()} is added.}

\item{shadow.args}{List. Arguments to \code{\link[ggfx:with_shadow]{ggfx::with_shadow()}}.}

\item{recompute}{Logical. If \code{TRUE}, use \link[ggplot2:lims]{lims()} to
replot the inset. Statistics, e.g. smoothing lines, will be
recomputed using only the data in the target area. If \code{FALSE}, use
\link[ggplot2:coord_cartesian]{coord_cartesian()} to replot the inset,
keeping all the data.}

\item{scale.inset}{Length 1 or 2 numeric. Normally, exactly the target area
is shown on the inset. Sometimes you may wish to rescale the plot in the
inset. Use 2 numbers to scale width and height separately.}

\item{proj.combine}{Logical. How to draw projection lines when more than
one polygon/map area is magnified? \code{FALSE} draws one set of projection
lines for each area. \code{TRUE} draws a single set of lines for all the areas.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
\code{geom_magnify()} creates a magnified inset of part of a ggplot. Optional
borders are drawn around the target and inset, along with projection lines
from one to the other. \code{from} gives the location of the target area,
and \code{to} gives the location of the inset. Usually, these are specified as
\code{c(xmin, xmax, ymin, ymax)}.
}
\details{
\subsection{Aesthetics}{

geom_magnify understand the following aesthetics (required aesthetics are in
bold):
\itemize{
\item \strong{from}
\item \strong{to}
}

\code{from} and \code{to} can be vectors of length 4, like \code{list(xmin, xmax, ymin, ymax)}.
These specify the bottom left and top right corners of the target area to
magnify, and the area for the magnified inset. The lists can optionally be
named: \code{list(xmin = 1, xmax = 2, ymin = 3, ymax = 4)}.

\strong{Note}: very early versions of ggmagnify used a
different order of coordinates: \code{list(xmin, ymin, xmax, ymax)}.

Alternatively, \code{from} can be:
\itemize{
\item A data frame of points with two columns for \code{x} and \code{y}, or a \code{\link[grid:grid.grob]{grid::grob()}}
object. Points within the grob region (a polygon spanned by the data frame)
will be magnified. Points should be
on the same scale as the data, with \code{default.units = "native"} in the grob.
\code{shape} will be ignored.
\item A logical vector. Points in the data where \code{from} is \code{TRUE} will be
surrounded by a rectangle, ellipse or outline.
}

Normally you'll set \code{from} and \code{to} in the call to \code{geom_magnify()}. You can
specify them as aesthetics, e.g. if you want different areas per facet. If
so, you may need to wrap them in a \code{\link[=list]{list()}} to make sure they are length one
per row of data. Only the first row per panel is used. (To magnify multiple
areas in one panel, use multiple calls to \code{geom_magnify()}.)
}

\subsection{Shapes}{

If \code{shape = "ellipse"} an elliptical area is magnified. This may not include
all points within the target area given by \code{from}.

If \code{shape = "outline"} then a convex hull will be drawn around points in
the target area. This only works if you are using \code{\link[=geom_point]{geom_point()}} or some other
geom with aesthetics \code{x} and \code{y}. If you are plotting a map, then
\code{"outline"} magnifies exactly the map features selected by \code{from}.
}

\subsection{Projection lines}{
\itemize{
\item \code{proj = "corresponding"} or \code{"facing"} draws projection lines from the
corners of the target to the corners of the inset.
\itemize{
\item \code{"corresponding"} always projects each corner of the target to the same
corner of the inset.
\item \code{"facing"} sometimes draws lines between facing corners, when this looks
cleaner.
\item For non-rectangular insets, \code{"facing"} and \code{"corresponding"} are the same.
}
\item \code{"single"} draws a single line from the midpoint of facing sides.
\item To draw no lines, set \code{proj.linetype = 0}.
}
}

\subsection{Limitations}{
\itemize{
\item \code{geom_magnify()} uses masks. This requires R version 4.1.0 or higher, and
a graphics device that supports masking. If you are using knitr, you may have
luck with the \code{ragg_png} device. If your device doesn't support masks,
only \code{shape = "rect"} will work, and the plot inset will not be clipped
to the panel area.
\item \code{geom_magnify()} uses dark magic to deal with faceting. It may break with
older, or newer, versions of ggplot2. If you don't need faceting, and want
your code to be robust to upgrades, set \code{options(ggmagnify.safe_mode = TRUE)}
to use slightly less magic.
\item By design, \code{geom_magnify()} replots the original plot using new limits. It
does not directly copy the target area pixels. The advantage is that you can
e.g. add axes, plot points at an appropriate size, zoom in on data that's
invisible in the main plot, or recompute derived graphics. If you want an
exact pixel-by-pixel copy, use a different tool.
\item \code{geom_magnify()} may break with discrete scales. This is a limitation in
ggplot2 for now.
\item Find a bug? Report it at \url{https://github.com/hughjonesd/ggmagnify/issues/}.
}
}

\subsection{\code{geom_magnify_tile()}}{

\code{geom_magnify_tile()} is a version of \code{geom_magnify()} which uses different
aesthetics. Set \code{x}, \code{width}, \code{y}, \code{height} and \code{to_x}, \code{to_width} \code{to_y},
\code{to_height} to specify the target and inset location.
}
}
\examples{
library(ggplot2)
ggp <- ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
         geom_point() + xlim(c(2, 6))
from <- list(2.5, 3.5, 6, 7)
to <- list(4, 6, 5, 7)

# Basic magnification
ggp + geom_magnify(from = from, to = to)

# Convex hull of points
ggp + geom_magnify(aes(from = Species == "setosa"), to = c(3, 5, 6, 8),
                   shape = "outline")

# Order matters

# `geom_magnify()` stores the plot when it is added to it:
ggp +
  scale_color_brewer() +
  geom_magnify(from = from, to = to)

# This will print the inset without the new scale:
ggp +
  geom_magnify(from = from, to = to) +
  scale_color_brewer()

# For more examples see https://github.com/hughjonesd/ggmagnify
}
