% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-magnify.R, R/geom-magnify-tile.R
\name{geom_magnify}
\alias{geom_magnify}
\alias{geom_magnify_tile}
\title{Magnified inset of a plot}
\usage{
geom_magnify(
  mapping = NULL,
  data = NULL,
  stat = StatMagnify,
  position = "identity",
  ...,
  from = NULL,
  to = NULL,
  expand = TRUE,
  aspect = c("free", "fixed"),
  axes = "",
  proj = c("facing", "corresponding", "single"),
  shadow = FALSE,
  linetype = 1,
  target.linetype = linetype,
  inset.linetype = linetype,
  proj.linetype = 2,
  alpha = 1,
  linewidth = 0.4,
  shape = c("rect", "ellipse"),
  plot = NULL,
  shadow.args = list(sigma = 5, colour = "grey40", x_offset = 5, y_offset = 5),
  recompute = FALSE,
  scale.inset = 1,
  na.rm = FALSE,
  inherit.aes = FALSE
)

geom_magnify_tile(
  mapping = NULL,
  data = NULL,
  stat = StatMagnifyTile,
  position = "identity",
  ...,
  expand = TRUE,
  aspect = c("free", "fixed"),
  axes = "",
  proj = "facing",
  shadow = FALSE,
  linetype = 1,
  target.linetype = linetype,
  inset.linetype = linetype,
  proj.linetype = 2,
  alpha = 1,
  linewidth = 0.4,
  shape = "rect",
  plot = NULL,
  shadow.args = list(sigma = 5, colour = "grey40", x_offset = 5, y_offset = 5),
  recompute = FALSE,
  scale.inset = 1,
  na.rm = FALSE,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm}{See e.g. \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{from}{Length 4 numeric: points \verb{x0, y0, x1, y1} of the target area to magnify.
Alternatively, pass in a data frame or \code{\link[grid:grid.grob]{grid::grob()}} object, see below.}

\item{to}{Length 4 numeric: points \verb{x0, y0, x1, y1} of the magnified inset.}

\item{expand}{Logical. Expand the limits of the target area by a small factor
in the inset plot. See \link[ggplot2:coord_cartesian]{coord_cartesian()}.}

\item{aspect}{String. \code{"fixed"} to fix the aspect ratio (overrides \code{ymax}).}

\item{axes}{String. Which axes to plot in the inset? \code{""}, \code{"x"}, \code{"y"} or
\code{"xy"}.}

\item{proj}{String. What style of projection lines to draw? \code{"facing"} (the
default), \code{"corresponding"} or \code{"single"}. Can be abbreviated. See below.}

\item{shadow}{Logical. Draw a shadow behind the inset plot? Requires the
"ggfx" package.}

\item{linetype, alpha, linewidth}{Linetype, alpha and linewidth for borders
and projection lines. Set \code{linetype = 0} for no lines.}

\item{target.linetype, inset.linetype, proj.linetype}{Linetypes
for specific components.}

\item{shape}{\code{"rect"} to magnify a rectangle. \code{"ellipse"} to magnify an ellipse.}

\item{plot}{Ggplot object to plot in the inset. If \code{NULL}, defaults to the
ggplot object to which \code{geom_magnify()} is added.}

\item{shadow.args}{List. Arguments to \code{\link[ggfx:with_shadow]{ggfx::with_shadow()}}.}

\item{recompute}{Logical. If \code{TRUE}, use \link[ggplot2:lims]{lims()} to
replot the inset. Statistics, e.g. smoothing lines, will be
recomputed using only the data in the target area. If \code{FALSE}, use
\link[ggplot2:coord_cartesian]{coord_cartesian()} to replot the inset,
keeping all the data.}

\item{scale.inset}{Length 1 or 2 numeric. Normally, exactly the target area
is shown on the inset. Sometimes you may wish to rescale the plot in the
inset. Use 2 numbers to scale width and height separately.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
\code{geom_magnify()} creates a magnified inset of part of a plot. Optional
borders are drawn around the target and inset, along with projection lines
from one to the other.
}
\details{
\subsection{Aesthetics}{

geom_magnify understand the following aesthetics (required aesthetics are in
bold):
\itemize{
\item \strong{xmin}
\item \strong{xmax}
\item \strong{ymin}
\item \strong{ymax}
\item \strong{to_xmin}
\item \strong{to_xmax}
\item \strong{to_ymin}
\item \strong{to_ymax}
\item colour
}

Normally you'll set these in the call to \code{geom_magnify()}, by specifying
\code{from} and \code{to}, but you can set them directly if you prefer to be explicit.
(\code{from} and \code{to} override \code{xmin} etc. and \code{to_xmin} etc. respectively.) Or
use the aesthetics to specify the magnification area within your data, e.g.
by facet.

\emph{Note:} as of today, the code only allows one magnification per call to
\code{geom_magnify()}. However, nothing stops you adding multiple calls to
\code{geom_magnify()}
}

\subsection{Projection lines}{

\code{proj = "corresponding"} or \code{"facing"} draws projection lines from the
corners of the target to the corners of the inset. \code{"corresponding"} always
projects each corner of the target to the same corner of the inset.
\code{"facing"} sometimes draws lines between facing corners, when this looks
cleaner. \code{"single"} draws a single line from the midpoint of facing sides.
\code{"none"} draws no lines.
}

\subsection{Arbitrary shapes}{

To magnify an arbitrary area, pass a \code{\link[grid:grid.grob]{grid::grob()}} or a data frame of x and y
points into \code{from}. Points should be on the same scale as the data,
with \code{default.units = "native"} in the grob. This is experimental.
(Well, everything is experimental. This is just \emph{more} experimental.)
}

\subsection{Limitations}{
\itemize{
\item \code{geom_magnify()} uses masks. This requires R version 4.1.0 or higher, and
a graphics device that supports masking. If you are using knitr, you may have
luck with the \code{ragg_png} device.
\item \code{geom_magnify()} uses dark magic to deal with faceting. It may break with
older (or newer!) versions of ggplot2.
\item By design, \code{geom_magnify()} replots the original plot using new limits. It
does not directly copy the target area pixels. The advantage is that you can
e.g. add axes, plot points at an appropriate size, zoom in on data that's
invisible in the main plot, or recompute derived graphics. If you want an
exact pixel-by-pixel copy, use a different tool.
\item \code{geom_magnify()} may break with discrete scales. This is a limitation in
ggplot2 for now.
}
}

\code{geom_magnify_tile()} is a version of \code{geom_magnify()} which uses different
aesthetics. Set \code{x}, \code{width}, \code{y}, \code{height} and \code{to_x}, \code{to_width} \code{to_y},
\code{to_height} to specify the target and inset location.
}
\examples{
library(ggplot2)
ggp <- ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
         geom_point() + xlim(c(2, 6))

# Basic magnification
ggp + geom_magnify(from = c(3, 6.5, 4, 7.5),
                     to = c(4, 5, 7, 6.5))

# Inset axes
ggp + geom_magnify(from = c(3, 6.5, 4, 7.5),
                     to = c(4, 5, 7, 6.5), axes = TRUE)

# Ellipse magnification
if (getRversion() >= 4.2) {
  ggp + geom_magnify(from = c(3, 6.5, 4, 7.5),
                     to = c(4, 5, 7, 6.5), shape = "ellipse")
}

# Shadow
if (requireNamespace("ggfx", quietly = TRUE)) {
  ggp + geom_magnify(from = c(3, 6.5, 4, 7.5),
                     to = c(4, 5, 7, 6.5), shadow = TRUE)
}

# Arbitrary shape using grid
setosas <- iris[iris$Species == "setosa", ]
setosa_hull <- grDevices::chull(setosas[, c("Sepal.Width", "Sepal.Length")])
setosa_hull <- setosas[setosa_hull, c("Sepal.Width", "Sepal.Length")]
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
       geom_point() + xlim(c(2, 5)) +
       geom_magnify(from = setosa_hull, to = c(3, 6, 5, 8))

# Order matters

# `geom_magnify()` stores the plot when it is added to it:
ggp +
  geom_smooth() +
  geom_magnify(from = c(3, 6.5, 4, 7.5),
               to = c(4, 5, 7, 6.5))

# This will print the inset without the smooth:
ggp +
  geom_magnify(from = c(3, 6.5, 4, 7.5),
               to = c(4, 5, 7, 6.5)) +
  geom_smooth()
}
