% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmagnify.R
\name{ggmagnify}
\alias{ggmagnify}
\title{Add a magnified inset plot to a ggplot object}
\usage{
ggmagnify(
  plot,
  xlim,
  ylim,
  inset_xlim,
  inset_ylim,
  zoom,
  border = TRUE,
  target = TRUE,
  proj = TRUE,
  shadow = FALSE,
  compose = TRUE,
  axes = FALSE,
  margin = if (axes) 10 else 0,
  linewidth = 0.5,
  linetype = 1,
  colour = "black",
  alpha = 1,
  inset_linewidth = linewidth,
  inset_linetype = linetype,
  inset_colour = colour,
  inset_alpha = alpha,
  proj_linewidth = linewidth,
  proj_linetype = "dashed",
  proj_colour = colour,
  proj_alpha = alpha,
  target_linewidth = linewidth,
  target_linetype = linetype,
  target_colour = colour,
  target_alpha = alpha,
  proj_style = c("facing", "corresponding"),
  shadow_args = list(sigma = 5, colour = "grey40", x_offset = 5, y_offset = 5),
  blank = inset_blanks(axes = axes)
)
}
\arguments{
\item{plot}{A ggplot object.}

\item{xlim, ylim}{Limits of the area to magnify. Length 2 numeric.}

\item{inset_xlim, inset_ylim}{Position of the inset in the main graph.
Length 2 numeric.}

\item{zoom}{How much to magnify the inset by. Numeric. Specify two numbers
for separate horizontal and vertical zoom. Optional: overrides \code{inset_xlim[2]}
and \code{inset_ylim[2]}.}

\item{border}{Logical. Draw a border around the inset?}

\item{target}{Logical. Draw a border around the target area?}

\item{proj}{Logical. Draw projection lines from the target area to the inset?}

\item{shadow}{Logical. Draw a shadow behind the inset? Requires
the \code{ggfx} package.}

\item{compose}{Logical. If \code{TRUE}, the new elements are added to the ggplot object.
If \code{FALSE}, they are returned in a \code{GgMagnify} object, see below.}

\item{axes}{Logical. Draw axes in the inset?}

\item{margin}{Plot margin of inset. Can be a single number in "pt"
units, a length 4 numeric (top, right, bottom, left), or a
\code{\link[ggplot2:element]{ggplot2::margin()}} object. Note that this is on the scale of the
inset plot, not the outer plot.}

\item{linewidth, linetype, colour, alpha}{Parameters for inset border, target border,
and projection lines.}

\item{inset_linewidth, inset_linetype, inset_colour, inset_alpha}{Parameters
for inset border.}

\item{proj_linewidth, proj_linetype, proj_colour, proj_alpha}{Parameters
for projection lines.}

\item{target_linewidth, target_linetype, target_colour, target_alpha}{Parameters for target border.}

\item{proj_style}{String. \code{"facing"} sometimes draws lines between
the nearest corners of the target and the inset.
\code{"corresponding"} always projects each corner of the target to the
corresponding corner of the inset. You can abbreviate these to \code{"f"}
or \code{"c"}.}

\item{shadow_args}{List of arguments to pass to \code{\link[ggfx:with_shadow]{ggfx::with_shadow()}}.}

\item{blank}{Character vector of theme elements to blank out in the inset.
Use \code{\link[=inset_blanks]{inset_blanks(elems, axes = axes)}} to add \code{elems}
to the default list.}
}
\value{
The modified \code{plot} if \code{compose} is \code{TRUE}. Otherwise, a \code{GgMagnify}
object.
}
\description{
\code{ggmagnify()} magnifies a target area of your plot and adds it as an inset on
top of the original plot. Optional projection lines and borders around the
target are drawn. If the \code{ggfx} package is installed, a drop shadow can be
added.
}
\details{
If \code{compose} is \code{FALSE}, the returned \code{GgMagnify} object includes the
following list components:
\itemize{
\item \code{inset}, a ggplot object representing the inset.
\item \code{border}, a layer representing the inset border.
\item \code{target}, a layer representing the target border.
\item \code{proj}, a layer representing the projection lines.
}

You can modify these, e.g. by adding themes to the inset. Call
\code{\link[=compose]{compose(ggm, plot)}} to add the object to the plot.

To create an inset outside the plot area, set \code{coord_cartesian(clip = "off")}
in the main plot.
\subsection{Limitations}{

It won't work with facets, or with non-cartesian coordinates.
}
}
\examples{
library(ggplot2)
ggp <- ggplot(diamonds, aes(carat, depth, color = cut)) + geom_point()

ggmagnify(ggp,
          xlim = c(1.5, 2.5), ylim = c(60, 65),
          inset_xlim = c(2, 5), inset_ylim = c(40, 55))

ggmagnify(ggp,
          xlim = c(1.5, 2.5), ylim = c(60, 65),
          inset_xlim = c(2, 5), inset_ylim = c(40, 55),
          colour = "red")

ggmagnify(ggp,
          xlim = c(1.5, 2.5), ylim = c(60, 65),
          inset_xlim = c(2, 5), inset_ylim = c(40, 55),
          axes = TRUE, border = FALSE)

if (requireNamespace("ggfx", quietly = TRUE)) {
  ggmagnify(ggp,
            xlim = c(1.5, 2.5), ylim = c(60, 65),
            inset_xlim = c(2, 5), inset_ylim = c(40, 55),
            shadow = TRUE)
}

```{r child = "man/Rmd/advanced-example.Rmd"}
```
}
