% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmagnify.R
\name{ggmagnify}
\alias{ggmagnify}
\title{Deprecated}
\usage{
ggmagnify(
  plot,
  xlim,
  ylim,
  inset_xlim,
  inset_ylim,
  zoom,
  border = TRUE,
  target = TRUE,
  proj = c("facing", "corresponding", "single", "none"),
  shadow = FALSE,
  axes = FALSE,
  margin = if (axes) 10 else 0,
  linewidth = 0.5,
  linetype = 1,
  colour = "black",
  alpha = 1,
  inset_linewidth = linewidth,
  inset_linetype = linetype,
  inset_colour = colour,
  inset_alpha = alpha,
  proj_linewidth = linewidth,
  proj_linetype = "dashed",
  proj_colour = colour,
  proj_alpha = alpha,
  target_linewidth = linewidth,
  target_linetype = linetype,
  target_colour = colour,
  target_alpha = alpha,
  shadow_args = list(sigma = 5, colour = "grey40", x_offset = 5, y_offset = 5),
  blank = inset_blanks(axes = axes),
  inset_expand = axes,
  inset_coord = ggplot2::coord_cartesian(xlim = xlim, ylim = ylim, expand = inset_expand)
)
}
\arguments{
\item{plot}{A ggplot object.}

\item{xlim, ylim}{Limits of the area to magnify. Length 2 numeric.}

\item{inset_xlim, inset_ylim}{Position of the inset in the main graph.
Length 2 numeric.}

\item{zoom}{How much to magnify the inset by.}

\item{border}{Logical. Draw a border around the inset?}

\item{target}{Logical. Draw a border around the target area?}

\item{proj}{String. What style of projection lines to draw?}

\item{shadow}{Logical. Draw a shadow behind the inset?}

\item{axes}{Logical. Draw axes in the inset?}

\item{margin}{Plot margin of inset.}

\item{linewidth, linetype, colour, alpha}{Parameters for inset border, target border,
and projection lines.}

\item{inset_linewidth, inset_linetype, inset_colour, inset_alpha}{Parameters
for inset border.}

\item{proj_linewidth, proj_linetype, proj_colour, proj_alpha}{Parameters
for projection lines.}

\item{target_linewidth, target_linetype, target_colour, target_alpha}{Parameters for target border.}

\item{shadow_args}{List of arguments to pass to \code{\link[ggfx:with_shadow]{ggfx::with_shadow()}}.}

\item{blank}{Character vector of theme elements to blank out in the inset.}

\item{inset_expand}{Logical. Expand the inset's view of the target?}

\item{inset_coord}{Result of a call to a \code{ggplot2::coord_} function to use
for the inset.}
}
\value{
A \code{GgMagnify} object.
}
\description{
\code{ggmagnify()} is deprecated. Use \code{\link[=geom_magnify]{geom_magnify()}} instead.
}
\keyword{internal}
\examples{
library(ggplot2)

ggp <- ggplot(ChickWeight, aes(Time, weight, group = Chick, color = Diet)) +
  geom_line()
ggmagnify(ggp,
          xlim = c(0, 5), ylim = c(30, 80),
          inset_xlim = c(0, 15), inset_ylim = c(200, 350))

}
