% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset-theme.R
\name{inset_blanks}
\alias{inset_blanks}
\title{Default elements to blank in the ggmagnify inset}
\usage{
inset_blanks(..., axes)
}
\arguments{
\item{...}{Character vector of extra elements to blank.}

\item{axes}{String. Which axes to plot in the inset? \code{""}, \code{"x"}, \code{"y"} or
\code{"xy"}.}
}
\value{
A character vector.
}
\description{
Default elements to blank in the ggmagnify inset
}
