% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset-theme.R
\name{inset_theme}
\alias{inset_theme}
\title{Create a theme suitable for an inset ggplot}
\usage{
inset_theme(
  blank = inset_blanks(axes = axes),
  axes,
  margin = if (axes) 10 else 0
)
}
\arguments{
\item{blank}{Character vector of extra elements to blank.
See \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{axes}{Logical: will the inset have axes?}

\item{margin}{Plot margin of inset.}
}
\value{
A ggplot theme object
}
\description{
Create a theme suitable for an inset ggplot
}
