% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset-theme.R
\name{inset_theme}
\alias{inset_theme}
\title{Create a theme suitable for an inset ggplot}
\usage{
inset_theme(
  blank = inset_blanks(axes = axes),
  axes,
  margin = if (axes == "") 0 else 8
)
}
\arguments{
\item{blank}{Character vector of theme elements to blank. See \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{axes}{String. Which axes to plot in the inset? \code{""}, \code{"x"}, \code{"y"} or
\code{"xy"}.}

\item{margin}{Margin around the plot. See \code{plot.margin} in \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A ggplot theme object
}
\description{
Use \code{inset_theme()} to add a suitable theme to a manually-created inset
plot.
}
\examples{
library(ggplot2)
ggp <- ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
         geom_point() + xlim(c(2, 6))
from <- list(2.5, 3.5, 6, 7)
to <- list(4, 6, 5, 7)

inset <- ggp + geom_density2d() + inset_theme(axes = "")
ggp + geom_magnify(from = from, to = to, plot = inset)
}
