% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset-theme.R
\name{inset_theme}
\alias{inset_theme}
\title{Create a theme suitable for an inset ggplot}
\usage{
inset_theme(
  blank = inset_blanks(axes = axes),
  axes,
  margin = if (axes == "") 0 else 8
)
}
\arguments{
\item{blank}{Character vector of theme elements to blank. See \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{axes}{String. Which axes to plot in the inset? \code{""}, \code{"x"}, \code{"y"} or
\code{"xy"}.}

\item{margin}{Margin around the plot. See \code{plot.margin} in \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A ggplot theme object
}
\description{
Create a theme suitable for an inset ggplot
}
