% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{+.GgMagnify}
\alias{+.GgMagnify}
\title{Add layer to a GgMagnify object}
\usage{
\method{+}{GgMagnify}(e1, e2)
}
\arguments{
\item{e1}{A GgMagnify object}

\item{e2}{A layer, theme, etc.}
}
\value{
The modified object
}
\description{
This adds the layer to \emph{both} the original plot, \emph{and} the inset.
}
\examples{

# Advanced usage

library(ggplot2)

booms <- ggplot(faithfuld, aes(waiting, eruptions)) +
         geom_contour_filled(aes(z = density)) +
         scale_fill_viridis_d(option = "B")

shadow_args <- list(
  colour = alpha("grey80", 0.8),
  x_offset = 0,
  y_offset = 0,
  sigma = 10
)

ggm <- ggmagnify(booms,
                 xlim = c(80, 92), ylim = c(4, 4.8),
                 inset_xlim = c(70, 94), inset_ylim = c(1.7, 3.3),
                 shadow = TRUE, shadow_args = shadow_args,
                 colour = "white")

# modify the inset only:
ggm$inset <- ggm$inset +
             geom_point(data = faithful, color = "red", fill = "white",
                        alpha = 0.7, size = 2, shape = "circle filled")

ggm

# modify the original plot only:

ggm$plot <- ggm$plot + scale_fill_grey()

ggm

# modify both:

ggm + scale_fill_viridis_d(option = "C")
}
